--- src/main/kotlin/domain/Dealer.kt
@@ -1,16 +1,0 @@
-package domain
-
-import domain.factory.DefaultCardsFactory
-
-class Dealer(
-    val deck: CardDeck = CardDeck(cardsFactory = DefaultCardsFactory)
-) {
-    init {
-        deck.shuffle()
-    }
-
-    fun giveCard(player: Player) {
-        val card = deck.pop()
-        player.receiveCard(card = card)
-    }
-}
